pyproject.toml
[tool.poetry]
name = "class-deck-backend"
version = "0.1.0"
description = "Write and share stories with card deck"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "class_deck_backend"}]

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"
django-environ = "^0.11.2"
factory-boy = "^3.3.0"
psycopg-binary = "^3.1.18"
pre-commit = "^3.7.0"
gunicorn = "^21.2.0"
pytest-django = "^4.8.0"
dj-rest-auth = "^5.1.0"
psycopg2 = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.3.0"
pytest-xdist = "^3.5.0"
drf-spectacular = "^0.27.2"
pyjwt = "^2.8.0"
pymemcache = "^4.0.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
pytest-django = "^4.8.0"
pytest-xdist = "^3.5.0"
isort = "^5.13.2"
black = "^24.3.0"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
